Backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
requests==2.32.3
modal-client==1.0.0
python-dotenv==1.0.1